Split NoPlaylists component.
